my_logbook/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.1
django-bootstrap-icons==0.7.5
django-cors-headers==3.11.0
django-oauth-toolkit==1.6.3
django-rest-framework==0.1.0
djangorestframework==3.13.1
filelock==3.4.2
idna==3.3
jwcrypto==1.0
oauthlib==3.1.1
platformdirs==2.4.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
virtualenv==20.13.0
wrapt==1.13.3
